refactor: fixed folder structure for another few helpers gf-382
