+++ apps/frontend/src/libs/hooks/use-pagination/libs/helpers/get-initial-page/get-initial-page.helper.ts
@@ -1,4 +1,4 @@
-import { FIRST_PAGE } from "../constants/constants.js";
+import { FIRST_PAGE } from "../../constants/constants.js";
 
 const getInitialPage = (
 	pageQueryParameter: null | string,
