+++ apps/frontend/src/libs/hooks/use-pagination/libs/helpers/helpers.ts
@@ -1,3 +1,3 @@
-export { getInitialPage } from "./get-initial-page.helper.js";
-export { getInitialPageSize } from "./get-initial-page-size.helper.js";
-export { calculateTotalPages } from "./total-pages.helper.js";
+export { getInitialPage } from "./get-initial-page/get-initial-page.helper.js";
+export { getInitialPageSize } from "./get-initial-page-size/get-initial-page-size.helper.js";
+export { calculateTotalPages } from "./total-pages/total-pages.helper.js";
