+++ apps/frontend/src/libs/hooks/use-pagination/libs/helpers/total-pages/total-pages.helper.ts
@@ -1,4 +1,4 @@
-import { FIRST_PAGE } from "../constants/constants.js";
+import { FIRST_PAGE } from "../../constants/constants.js";
 
 const calculateTotalPages = (pageSize: number, totalItems: number): number =>
 	Math.max(FIRST_PAGE, Math.ceil(totalItems / pageSize));
