+++ apps/frontend/src/pages/access-management/libs/components/group-form/libs/helpers/get-permission-options/get-permission-options.helper.ts
