+++ apps/frontend/src/pages/project-access-management/libs/components/project-group-form/libs/helpers/get-permission-options/get-permission-options.helper.ts
